moving files to proper folders
