+++ app/src/main/java/com/lucasmontano/carreiranomundomobile/collections/HabitListAdapter.kt
@@ -5,6 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import com.lucasmontano.carreiranomundomobile.collections.model.HabitItem
 import com.lucasmontano.carreiranomundomobile.databinding.HabitItemBinding
